requirements.txt
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
commonmark==0.9.1
constantly==15.1.0
cryptography==38.0.4
cssselect==1.2.0
dnspython==2.2.1
et-xmlfile==1.1.0
filelock==3.9.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
jmespath==1.0.1
loguru==0.6.0
lxml==4.9.2
numpy==1.24.1
openai==0.25.0
openpyxl==3.0.10
packaging==22.0
pandas==1.5.2
pandas-stubs==1.5.2.221213
parsel==1.7.0
pinecone-client==2.0.13
Protego==0.2.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyDispatcher==2.0.6
Pygments==2.14.0
pyOpenSSL==22.1.0
python-dateutil==2.8.2
pytz==2022.7
PyYAML==6.0
queuelib==1.6.2
requests==2.28.1
requests-file==1.5.1
Scrapy==2.7.1
service-identity==21.1.0
six==1.16.0
tldextract==3.4.0
tqdm==4.64.1
Twisted==22.10.0
types-pytz==2022.7.0.0
typing_extensions==4.4.0
urllib3==1.26.13
w3lib==2.1.1
zope.interface==5.5.2
